Move classes from utility to scGIST
